Move NodeApi subclasses to node_api directory
